+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Bug fix:
 - Ensure constraints and properties are string-keyed so that they match regardless of which are used (#33)
 - Be more permissive with the situations where `Rspec::Determinator` can be used (#34)
+- Swallow not found errors from routemaster so determinator isn't too shouty, allow them to be tracked. (#35)
 
 # v0.10.0 (2017-09-15)